pyproject.toml
[tool.poetry]
name = "nonebot-plugin-jrrp3"
version = "3.3.2"
description = "更加现代化的 NoneBot2 每日人品插件，支持查询今日、本周、本月和历史平均人品，自定义运势，以及数据持久化存储。"
authors = ["Rene8028", "UKMeng <ukmengse@gmail.com>", "GT610 <myddz1005@outlook.com>"]
readme = "README.md"
packages = [{include = "nonebot_plugin_jrrp3"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
nonebot2 = ">=2.3.0,<3.0.0"
nonebot-plugin-alconna = ">=0.50.0"
nonebot-plugin-localstore = ">=0.6.0"
pyyaml = ">=6.0"

[tool.poetry.group.dev.dependencies]
nonebot2 = {extras = ["fastapi", "httpx", "webscokets"], version = ">=2.1.0"}
nonebot-adapter-onebot = ">=2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = [ "nonebot_plugin_jrrp3"]
plugin_dirs = []
builtin_plugins = ["echo"]